Sussythings Beta 0.6.0 NEW INSTALL METHOD
Simplifies a bit the setup for the mod and the complexity of the README (install instructions) by creating folders automatically pressing INSTALL.bat into Minecraft, surpassing the need for self-organized folders by the customer.  This process also removes the need to quit and reload the installation for the mod to work. I added sections to the README to make sure it looks better and feels much easier to read. Please test this new installation method and write your thoughts in my discord server "Sussythings mod" link: https://discord.gg/Rwq292fxam